requirements.txt
langchain_core
langchain_community
pypdf
streamlit
bs4
langchain-huggingface
